Updated code changes as per review comments
