+++ src/LCT.PackageIdentifier/Program.cs
@@ -75,10 +75,10 @@ namespace LCT.PackageIdentifier
 
             // Validate application settings
             await ValidateAppsettingsFile(appSettings, projectReleases);
-            string listOfInlude = DisplayIncludeFiles(appSettings);
-            string listOfExclude = DisplayExcludeFiles(appSettings);
-            string listOfExcludeComponents = DisplayExcludeComponents(appSettings);            
-            string listOfInternalRepoList = GetInternalRepolist(appSettings);
+            string listOfInlude = DisplayInformation.DisplayIncludeFiles(appSettings);
+            string listOfExclude = DisplayInformation.DisplayExcludeFiles(appSettings);
+            string listOfExcludeComponents = DisplayInformation.DisplayExcludeComponents(appSettings);            
+            string listOfInternalRepoList = DisplayInformation.GetInternalRepolist(appSettings);
 
             Logger.Logger.Log(null, Level.Notice, $"Input Parameters used in Package Identifier:\n\t" +
                 $"CaToolVersion\t\t --> {caToolInformation.CatoolVersion}\n\t" +
@@ -158,107 +158,7 @@ namespace LCT.PackageIdentifier
                 environmentHelper=new EnvironmentHelper();
                 environmentHelper.CallEnvironmentExit(-1);
             }
-        }
-        private static string DisplayIncludeFiles(CommonAppSettings appSettings)
-        {
-            string totalString = string.Empty;
-            var includeMappings = new Dictionary<string, Func<IEnumerable<string>>>(StringComparer.OrdinalIgnoreCase)
-    {
-        { "NPM", () => appSettings.Npm.Include },
-        { "NUGET", () => appSettings.Nuget.Include },
-        { "MAVEN", () => appSettings.Maven.Include },
-        { "DEBIAN", () => appSettings.Debian.Include },
-        { "POETRY", () => appSettings.Poetry.Include },
-        { "CONAN", () => appSettings.Conan.Include },
-        { "ALPINE", () => appSettings.Alpine.Include }
-    };
-
-            if (includeMappings.TryGetValue(appSettings.ProjectType, out var getIncludeList))
-            {
-                var includeList = getIncludeList();
-                if (includeList != null)
-                {
-                    totalString = string.Join(",", includeList);
-                }
-            }
-            else
-            {
-                Logger.Error($"Invalid ProjectType - {appSettings.ProjectType}");
-            }
-
-            return totalString;
-        }
-        private static string DisplayExcludeFiles(CommonAppSettings appSettings)
-        {
-            string totalString = string.Empty;
-            var excludeMappings = new Dictionary<string, Func<IEnumerable<string>>>(StringComparer.OrdinalIgnoreCase)
-    {
-        { "NPM", () => appSettings.Npm.Exclude },
-        { "NUGET", () => appSettings.Nuget.Exclude },
-        { "MAVEN", () => appSettings.Maven.Exclude },
-        { "DEBIAN", () => appSettings.Debian.Exclude },
-        { "POETRY", () => appSettings.Poetry.Exclude },
-        { "CONAN", () => appSettings.Conan.Exclude },
-        { "ALPINE", () => appSettings.Alpine.Exclude }
-    };
-
-            if (excludeMappings.TryGetValue(appSettings.ProjectType, out var getExcludeList))
-            {
-                var excludeList = getExcludeList();
-                if (excludeList != null)
-                {
-                    totalString = string.Join(",", excludeList);
-                }
-            }
-            else
-            {
-                Logger.Error($"Invalid ProjectType - {appSettings.ProjectType}");
-            }
-
-            return totalString;
-        }
-
-        private static string DisplayExcludeComponents(CommonAppSettings appSettings)
-        {
-
-            string totalString = string.Empty;
-            if (appSettings.SW360.ExcludeComponents != null)
-            {
-                totalString = string.Join(",", appSettings.SW360.ExcludeComponents?.ToList());
-            }
-            return totalString;
-        }
-
-        private static string GetInternalRepolist(CommonAppSettings appSettings)
-        {
-            string listOfInternalRepoList = string.Empty;
-
-            var repoMapping = new Dictionary<string, Func<IEnumerable<string>>>(StringComparer.OrdinalIgnoreCase)
-        {
-        { "NPM", () => appSettings.Npm?.Artifactory.InternalRepos },
-        { "NUGET", () => appSettings.Nuget?.Artifactory.InternalRepos },
-        { "MAVEN", () => appSettings.Maven?.Artifactory.InternalRepos },
-        { "DEBIAN", () => appSettings.Debian?.Artifactory.InternalRepos },
-        { "POETRY", () => appSettings.Poetry?.Artifactory.InternalRepos },
-        { "CONAN", () => appSettings.Conan?.Artifactory.InternalRepos },
-        { "ALPINE", () => appSettings.Alpine?.Artifactory.InternalRepos }
-        };
-
-            if (repoMapping.TryGetValue(appSettings.ProjectType, out var getRepos))
-            {
-                var repos = getRepos();
-                if (repos != null)
-                {
-                    listOfInternalRepoList = string.Join(",", repos);
-                }
-            }
-            else
-            {
-                Logger.Error($"Invalid ProjectType - {appSettings.ProjectType}");
-            }
-
-            return listOfInternalRepoList;
-        }
+        }       
 
         private static string LogFolderInitialisation(CommonAppSettings appSettings)
         {
